% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{get_data}
\alias{get_data}
\title{Get bikeshare data for a particular URL}
\usage{
get_data(url)
}
\arguments{
\item{url}{URL for data, get from \code{\link{feeds_urls}}}
}
\value{
dataframe of data
}
\description{
Get bikeshare data for a particular URL
}
\examples{
if(interactive()) {
feeds_urls() \%>\%
dplyr::filter(name == "system_information") \%>\%
dplyr::pull("url") \%>\%
get_data()
}
}
