% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{oos_metrics}
\alias{oos_metrics}
\title{Generate metrics for a model}
\usage{
oos_metrics(real, pred)
}
\arguments{
\item{real}{vector of real}

\item{pred}{vector of predictions}
}
\value{
tibble of goodness-of-fit metrics
}
\description{
Generate metrics for a model
}
\examples{
oos_metrics(1:3, 4:6)
}
