% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{print.bike_data}
\alias{print.bike_data}
\title{Print bike data}
\usage{
\method{print}{bike_data}(x)
}
\arguments{
\item{x}{object of class "bike_data"}
}
\value{
dataframe of data, invisibly
}
\description{
Print bike data
}
\examples{
if(interactive()) {
feeds_urls() \%>\%
dplyr::filter(name == "system_information") \%>\%
dplyr::pull("url") \%>\%
get_data()
}
}
